+++ tests/test_fornax_base.py
@@ -1,3 +1,5 @@
+# Copyright 2026, University of Maryland, All Rights Reserved
+
 import sys
 import os
 import shutil