+++ changes.md
@@ -1,4 +1,16 @@
+<!--
+Copyright 2026, University of Maryland, All Rights Reserved
+-->
+
 # Latest Changes
+
+## 26.0914
+- Expose the MAST and IRSA environments in the Jupyterlab launcher.
+- Upgrade the default python to 3.13.
+- Update fornax-labextension to 0.1.22: Change the label Introduction to Tutorials.
+- Add s3archive script to help efficiently archive data to ~/s3-storage.
+- Bug fixes and enhancements:
+  - Firefly now uses IRSA's cloud backend.
 
 ## 26.0814
 - Hotfix for the missing /opt/support-data symlink.